Mezuniyet projesi/requirements.txt
discord.py
sounddevice
numpy
scipy
SpeechRecognition
